Moved cache queues into a separate source file
